All hardware devices are moved outside CxbxKrnl, towards devices (some into devices\video)
