+++ src/devices/SMCDevice.cpp
@@ -41,8 +41,9 @@ namespace xboxkrnl
 #include <xboxkrnl/xboxkrnl.h> // For xbox.h:AV_PACK_HDTV
 };
 
-#include "CxbxKrnl.h"
-#include "EmuShared.h"
+#include "CxbxKrnl\CxbxKrnl.h"
+#include "CxbxKrnl\EmuShared.h"
+
 #include "SMCDevice.h" // For SMCDevice
 #include "LED.h"
