+++ src/devices/SMDevice.cpp
