+++ src/devices/video/nv2a_int.h
